requirements.txt
Flask>=1.0,<3.0
google-cloud-storage
